# Remove boxed primitives from aggregations